Add plugin tagbar
Tagbar is a Vim plugin that provides an easy way to browse the tags
of the current file and get an overview of its structure.
It does this by creating a sidebar that displays the ctags-generated
tags of the current file, ordered by their scope.
This means that for example methods in C++ are displayed under the class
they are defined in.

Signed-off-by: xlongfeng <xlongfeng@126.com>